% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcars.R
\name{lcars_colors}
\alias{lcars_colors}
\alias{lcars_2357}
\alias{lcars_2369}
\alias{lcars_2375}
\alias{lcars_2379}
\title{Hex colors from LCARS color names}
\usage{
lcars_colors(...)

lcars_2357(...)

lcars_2369(...)

lcars_2375(...)

lcars_2379(...)
}
\arguments{
\item{...}{character, LCARS color names.}
}
\description{
Obtain hex colors from standard LCARS color names.
}
\details{
These functions return the hex colors for LCARS color names. If no argument is provided, the functions return their respective LCARS color set.
These functions correspond to LCARS color palettes that contain named colors.
Other predefined LCARS color palettes are available but do not have names for each color.
You can see all LCARS palettes with \code{\link{lcars_pals}}.
}
\examples{
lcars_colors()
lcars_2357()
lcars_colors("rust", "danub")
}
\seealso{
\code{\link{lcars_pals}}
}
