% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trekcolors.R
\docType{package}
\name{trekcolors}
\alias{trekcolors-package}
\alias{trekcolors}
\title{trekcolors: Star Trek Color Palettes}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a dataset of predefined color palettes based on the Star Trek science fiction series, associated color palette functions, and additional functions for generating customized palettes that are on theme. The package also offers functions for applying the palettes to plots made using the 'ggplot2' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/leonawicz/trekcolors}
  \item Report bugs at \url{https://github.com/leonawicz/trekcolors/issues}
}

}
\author{
\strong{Maintainer}: Matthew Leonawicz \email{rpkgs@pm.me} (\href{https://orcid.org/0000-0001-9452-2771}{ORCID})

}
