% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trekcolors.R
\name{trekpal}
\alias{trekpal}
\title{Named Star Trek color palette}
\usage{
trekpal(id)
}
\arguments{
\item{id}{character, name of predefined palette. If missing, return all available palette names.}
}
\value{
character vector of hex colors or palette names
}
\description{
Return a named Star Trek color palette from the \code{trekpals} dataset.
}
\examples{
trekpal("lcars2357")
# leave id blank to list available names:
trekpal()
}
